añadiendio restricciones

+++ MVMood/views/publicaciones_listado.php
@@ -30,7 +30,7 @@ include 'header.php';
         	</div>
         	<div class="post-content"><?= nl2br(htmlspecialchars($p['contenido'])) ?></div>
         	<div class="post-actions">
-            	<?php if ($p['idUsuario'] == $_SESSION['id'] || ACL::puede('publicaciones.eliminar')): ?>
+            	<?php if ($_SESSION['rol']=='admin' || $p['idUsuario'] == $_SESSION['id'] and ACL::puede('publicaciones.eliminar') ): ?>
                 	<a href="index.php?controller=Publicaciones&action=editar_publicacion&id=<?= $p['id'] ?>" class="action-btn">✏️ Editar</a>
                 	<a href="index.php?controller=Publicaciones&action=eliminar&id=<?= $p['id'] ?>" class="action-btn" onclick="return confirm('¿Seguro que quieres eliminar esta publicación?')">🗑️ Eliminar</a>
             	<?php endif; ?>
